feat: 实现 Gmail OAuth2 认证和邮件同步功能
- 添加 OAuth2 认证服务和处理器，支持 Google 和 Microsoft 账户
- 实现 Gmail API 邮件同步功能，支持增量同步
- 修复 SyncLog 模型数据类型问题，解决 500 错误
- 添加前端 OAuth2 认证组件和回调页面
- 实现 Token 自动刷新服务，提升用户体验
- 完善错误处理和日志记录，移除敏感信息
- 更新配置文件和路由，支持新的认证流程

核心功能:
- OAuth2 授权 URL 生成和状态验证
- 授权码交换和用户信息获取
- 访问令牌自动刷新机制
- Gmail 邮件增量同步
- 前端授权流程和状态管理

技术改进:
- 修复数据库模型类型不匹配问题
- 优化凭证加密和解析逻辑
- 清理调试日志，确保生产环境安全
- 添加完整的错误处理和用户反馈

测试结果:
- Gmail 账户成功同步 11 封邮件
- OAuth2 授权流程正常工作
- Token 刷新服务运行正常
- API 接口返回正确数据

+++ backend/internal/model/sync_log.go
@@ -14,14 +14,14 @@ type SyncLog struct {
 	Status   string `gorm:"size:20;not null;index" json:"status"` // running/success/failed
 
 	// 统计信息
-	EmailsFetched int `gorm:"default:0" json:"emails_fetched"`
-	EmailsNew     int `gorm:"default:0" json:"emails_new"`
-	EmailsUpdated int `gorm:"default:0" json:"emails_updated"`
+	EmailsFetched int64 `gorm:"default:0" json:"emails_fetched"`
+	EmailsNew     int64 `gorm:"default:0" json:"emails_new"`
+	EmailsUpdated int64 `gorm:"default:0" json:"emails_updated"`
 
 	// 时间信息
 	StartedAt   time.Time  `gorm:"not null;index:idx_started_at,sort:desc" json:"started_at"`
 	CompletedAt *time.Time `json:"completed_at"`
-	DurationMs  int        `json:"duration_ms"`
+	DurationMs  int64      `json:"duration_ms"`
 
 	// 错误信息
 	ErrorMessage string `gorm:"type:text" json:"error_message"`
